Get names from typedefs for anonymous structs
In C, an anonymous struct can be named using a typedef. Prior to this
commit, dwarf2json ignores typedefs and generates unnamed user_types for
anonymous structs. This makes it difficult to refer to those types.

This commit uses a typedef's name to refer to a user_type when
dwarf2json encounters an anonymous struct instead of using a generated
name.